computeBalanceChange to utils
